+++ Practica2/Models/DOCTOR.cs
@@ -21,7 +21,7 @@ namespace Practica2.Models
             this.ESPECIALIDAD_DOCTOR = new HashSet<ESPECIALIDAD_DOCTOR>();
         }
     
-        public int dpi { get; set; }
+        public long dpi { get; set; }
         public string nombre { get; set; }
         public string direccion { get; set; }
         public int telefono { get; set; }
